apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
redis==5.0.7
boto3==1.34.161
python-dotenv==1.0.1
anthropic==0.34.2
datasets==2.19.0
requests==2.32.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
